requirements.txt
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
faker==25.2.0
google-cloud-storage==2.18.2
